requirements-dev.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.7
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
Pygments==2.2.0
python-decouple==3.1
pytz==2018.5
requests==2.19.1
reverse==0.1.0
simplejson==3.16.0
static3==0.7.0
uritemplate==3.0.0
urllib3==1.23
wincertstore==0.2
